+++ NEWS.md
@@ -3,6 +3,7 @@
 New features:
 
 * Added support for generalised least squares models (class "gls")
+* Added support for beta regression models (class "betareg")
 
 Bugs fixed:
 
@@ -10,7 +11,7 @@ Bugs fixed:
 involving multi-coefficient terms (e.g. factors)
 * Function 'xNam' does not generate correct term names for factors when the
 model intercept is suppressed
-* Function 'R2' with argument 'pred = TRUE' does not work for models where any
+* Function 'R2' with argument 'pred = TRUE' throws an error for models where any
 weights = 0
 
 ## semEff 0.1.0